Setup helpers, migrations, & test harness for testing time jumps
During migration we have to get the current time from the chain itself,
without making assumptions about the relationship between the local
machine time and the time on chain. This allows us to run tests repeatedly
on the same long-running instance of a ganache-cli mainnet fork, w/o
having to restart the local testnet each time to execute the tests.

Note that this is depdendent on the loading order of Truffle. Assumes that
the migrations load before the tests, so the static property that saves the
network timestamp will be preserved.